requirements.txt
# Web Framework
fastapi==0.115.13
uvicorn[standard]==0.34.3
websockets==15.0.1

# UI Framework
nicegui==2.20.0

# Database
sqlalchemy==2.0.41
# drizzle-orm==0.29.0  # Will add later when needed

# AI and Language Models
langchain>=0.1.0
langchain-mistralai==0.2.10
mistralai==1.8.2

# Voice Synthesis
elevenlabs==0.2.26

# Data Validation
pydantic==2.11.7

# Authentication
PyJWT==2.10.1

# Async Support
aiofiles==24.1.0
httpx==0.28.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Utilities
python-dotenv==1.1.0

# Optional: For production
gunicorn==21.2.0 